requirements.txt
asgiref==3.6.0
Django==4.1.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
joblib==1.2.0
numpy==1.24.1
pandas==1.5.3
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tzdata==2022.7
xgboost==1.7.3
